1) In LeavesView.m, if you set a breakpoint in getImages, you'll notice that method is called twice when a LeavesViewController subclass is first loaded, when once would be sufficient.  This can be fixed by moving the "self.currentPageIndex = 0;" line from reloadData to layoutSubviews.
2) Also added an OverlayView class which works as you described in
June (below).  It works from both initWithFrame:nibBundle and from a
NIB.  When a series of touch events starts within the targetWidth of
either side, it forwards all of the events in the series to the
LeavesView, otherwise it will forward that series of events to the
UIViewController.  There's an example that registers a delegate to
update the overlay when the page is turned, as you suggested.